gsheets_auth/requirements.txt
fastapi
uvicorn
pydantic
pydantic[email]
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
passlib[bcrypt]
bcrypt
certifi
python-dotenv
